backend/requirements.txt
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.6
daphne==4.2.1
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
msgpack==1.1.1
pillow==11.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.1.0
python-dotenv==1.1.1
PyYAML==6.0.2
qrcode==8.2
redis==6.4.0
referencing==0.36.2
rpds-py==0.27.1
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
zope.interface==7.2
dj-database-url==2.3.0
psycopg2-binary==2.9.9
sentry-sdk[django]==2.17.0
pytest-asyncio==0.23.6
